requirements.txt
Pillow==12.0.0
requests==2.32.5
pytest==9.0.1
tomli>=2.0.1; python_version < "3.11"
